refactored the existing code (fixed issue with driver)
