+++ src/main/java/pages/BasePage.java
@@ -1,16 +1,13 @@
 package pages;
 
-import static drivermanager.CustomWebDriverManager.getDriverManagerInstance;
+import static drivermanager.CustomWebDriverManager.getDriverInstance;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
 public class BasePage {
-    public WebDriver driver() {
-        return getDriverManagerInstance().getDriver();
-    }
+
     public WebElement getElement(By elem) {
-        return driver().findElement(elem);
+        return getDriverInstance().findElement(elem);
     }
 }
