+++ src/main/resources/stories/Login.story
@@ -1,5 +1,6 @@
 Scenario: 1.Check that the user can log in to ReportPortal
 Meta:
+@to_run
 Given User opens the 'http://localhost:8080/ui/#login' web site
 When User logs in with the next data:
 | login      | password |
